// bounded-contexts/build-infra/domain/enums/azure-resources.enum.ts
export enum AzureResourceEnum {
    VPC = "azurerm_virtual_network",
    VPC_SUB = "azurerm_subnet",
    VPC_SG = "azurerm_network_security_group",
    VM = "azurerm_virtual_machine",
    STORAGE = "azurerm_storage_account",
    MS_SQL_SERVER = "azurerm_mssql_server",
    MS_SQL_DB = "azurerm_mssql_database",
    COSMOS_DB_ACC = "azurerm_cosmosdb_account",
    COSMOS_DB = "azurerm_cosmosdb_database",
    COSMOS_DB_CONTAINER = "azurerm_cosmosdb_container",
    REDIS = "azurerm_redis_cache",
    FW_RULE = "azurerm_network_security_rule"

}